primetoaster.sh,toaster-launch.sh: sqlite moved
Toaster moved the location of toaster.sqlite from the
BUILDDIR to the BUILDDIR/.. after morty.  This patchset
deals with that move for both priming and launching.

+++ primetoaster.sh
@@ -33,6 +33,14 @@ cd $WORKDIR
 . $POKYDIR/bitbake/bin/toaster start
 
 . $POKYDIR/bitbake/bin/toaster stop
-
-mv toaster.sqlite $WORKDIR || exit 1
+T_DB_FILE="toaster.sqlite"
+T_DB=`find $HOME -iname $T_DB_FILE`
+if [ $(dirname $T_DB) != $(readlink -f  $WORKDIR) ]; then
+    echo "Moving $T_DB/$T_DB_FILE to $WORKDIR"
+    mv ${T_DB} $WORKDIR || exit 1
+fi
+if [ ! -f $WORKDIR/$T_DB_FILE ]; then
+    echo "$WORKDIR/$T_DB_FILE not found, erroring"
+    exit 1
+fi
 rm $WORKDIR/build -rf || exit 1
